rename packages to plural troughout
