+++ src/main/java/be/syntra/devshop/DevshopFront/models/SaveStatus.java
@@ -1,4 +1,4 @@
-package be.syntra.devshop.DevshopFront.model;
+package be.syntra.devshop.DevshopFront.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
